requirements.txt
# FastAPI y servidor ASGI
fastapi==0.111.1
uvicorn[standard]==0.30.3
gunicorn==22.0.0

# Pydantic y configuración
pydantic==2.11.4
pydantic-settings==2.9.1
python-dotenv==1.1.0

# Azure SDK
azure-identity==1.17.0
azure-core>=1.34.0

# Google Ads SDK
google-ads>=25.0.0
google-auth-oauthlib>=0.5.1
google-api-python-client>=2.0.0

# Facebook Business SDK (para Meta Ads)
facebook-business>=19.0.0

# TikTok Business API SDK
tiktok-business-api-sdk>=0.2.6

# HTTP y utilidades
requests==2.32.3
typing-extensions>=4.12.2
typing-inspection>=0.4.0

# Criptografía
cryptography==43.0.3
cffi>=1.14
pycparser>=2.22

# MSAL (dependencia transitiva o para otros flujos)
msal>=1.24.0
msal-extensions>=0.3.0
PyJWT[crypto]<3,>=1.0.0

# Otras utilidades
annotated-types==0.7.0
certifi>=2017.4.17
charset-normalizer<4,>=2
idna<4,>=2.5
six>=1.11.0
urllib3<3,>=1.21.1
protobuf>=3.19.0
PyYAML>=5.4 # Añadido para GoogleAdsClient si se usa carga desde YAML, o para TikTok SDK si lo requiere.